Move a component to a helper
